BUG/MINOR: use the STRLCPY() macro instead of the strncpy() function
When using the gcc compiler (version 10), the following warning is printed:
  warning: 'strncpy' specified bound 8192 equals destination size [-Wstringop-truncation]

This actually denotes two bugs in the source, where it was possible to copy
the source string to the destination without terminating the destination
string with null byte in case the length of the source string is greater
than or equal to the size of the destination string.

As some systems do not have the strlcpy() function that can be used instead
of strncpy(), its operation is simulated via the STRLCPY() macro.

Version of the program changed to v1.0.1.